requirements.txt
faiss-cpu==1.8.0.post1
fastjsonschema==2.20.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Limiter==3.7.0
Flask-PyMongo==2.3.0
google-ai-generativelanguage==0.6.5
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.63.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
langchain==0.2.15
langchain-community==0.2.14
langchain-core==0.2.36
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.2
Markdown==3.6
markdown2==2.5.0
MarkupSafe==2.1.5
numpy==1.26.4
pymongo==4.8.0
